# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Result": {
				"properties": {
					"link": {
						"type": "string"
					}
				},
				"required": [
					"link"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Error": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "kawaii-image-api",
		"version": "0.1",
		"description": "REST API to serve different images or gifs from animes",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "eme22"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/sfw/neko": {
			"get": {
				"operationId": "GetRandomNeko",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								}
							}
						}
					},
					"404": {
						"description": "Theres no images in this endpoint",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"description": "Retrieves a random Neko link",
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": []
			}
		},
		"/sfw/neko/{id}": {
			"get": {
				"operationId": "GetNeko",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Result"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Image not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/sfw/kiss": {
			"get": {
				"operationId": "GetRandomKiss",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								}
							}
						}
					},
					"404": {
						"description": "Theres no images in this endpoint",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"description": "Retrieves a random Kissing GIF link",
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": []
			}
		},
		"/sfw/kiss/{id}": {
			"get": {
				"operationId": "GetKiss",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Result"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Image not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/sfw/waifu": {
			"get": {
				"operationId": "GetRandomWaifu",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								}
							}
						}
					},
					"404": {
						"description": "Theres no images in this endpoint",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"description": "Retrieves a random Waifu link",
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": []
			}
		},
		"/sfw/waifu/{id}": {
			"get": {
				"operationId": "GetWaifu",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Result"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Image not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/sfw/cry": {
			"get": {
				"operationId": "GetRandomCry",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								}
							}
						}
					},
					"404": {
						"description": "Theres no images in this endpoint",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"description": "Retrieves a random Crying GIF link",
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": []
			}
		},
		"/sfw/cry/{id}": {
			"get": {
				"operationId": "GetCry",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Result"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Image not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/sfw/hug": {
			"get": {
				"operationId": "GetRandomHug",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								}
							}
						}
					},
					"404": {
						"description": "Theres no images in this endpoint",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"description": "Retrieves a random Huging GIF link",
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": []
			}
		},
		"/sfw/hug/{id}": {
			"get": {
				"operationId": "GetHug",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Result"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Image not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/sfw/lick": {
			"get": {
				"operationId": "GetRandomLick",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								}
							}
						}
					},
					"404": {
						"description": "Theres no images in this endpoint",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"description": "Retrieves a random Licking GIF link",
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": []
			}
		},
		"/sfw/lick/{id}": {
			"get": {
				"operationId": "GetLick",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Result"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Image not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/sfw/pat": {
			"get": {
				"operationId": "GetRandomPat",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								}
							}
						}
					},
					"404": {
						"description": "Theres no images in this endpoint",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"description": "Retrieves a random Patting GIF link",
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": []
			}
		},
		"/sfw/pat/{id}": {
			"get": {
				"operationId": "GetPat",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Result"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Image not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/sfw/smug": {
			"get": {
				"operationId": "GetRandomSmug",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								}
							}
						}
					},
					"404": {
						"description": "Theres no images in this endpoint",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"description": "Retrieves a random Smugging GIF link",
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}